feat(ms-word-addin): print button
